--- a/src/bench/checkqueue.cpp
+++ b/src/bench/checkqueue.cpp
@@ -16,7 +16,7 @@
 
 static const size_t BATCHES = 101;
 static const size_t BATCH_SIZE = 30;
-static const int PREVECTOR_SIZE = 28;
+static const int PREVECTOR_SIZE = 36;
 static const unsigned int QUEUE_BATCH_SIZE = 128;
 
 // This Benchmark tests the CheckQueue with a slightly realistic workload,