… src/bench/checkblock.cpp
… src/bench/crypto_hash.cpp
… src/bench/mempool_stress.cpp
… src/bench/prevector.cpp
… src/bench/rpc_blockchain.cpp
… src/bench/xor.cpp
… src/obfuscation.h
… src/streams.cpp
… src/streams.h
… src/test/crypto_tests.cpp
… src/test/dbwrapper_tests.cpp
… src/test/fuzz/autofile.cpp
… src/test/fuzz/buffered_file.cpp
… src/test/fuzz/integer.cpp
… src/test/hash_tests.cpp
… src/test/script_tests.cpp
… src/test/streams_tests.cpp
… src/test/transaction_tests.cpp
… src/test/util/setup_common.cpp
… src/test/util/setup_common.h
… src/test/validation_flush_tests.cpp
… src/util/hasher.cpp
… src/util/hasher.h