During the last Core Dev meeting, it was proposed to create a tracking PR aggregating the individual IBD optimizations - to illustrate how these changes contribute to the broader performance improvement efforts.

## Summary: **18%** full IBD speedup
We don't have many low-hanging fruits anymore, but big speed improvements can also be achieved by many small, focused changes.
Many optimization opportunities are hiding in consensus critical code - this tracking PR provides justification for why those should also be considered.
The unmerged changes here collectively achieve a **~18%** speedup for full IBD (measured by multiple real runs until 886'000 blocks using 5GiB in-memory cache): from **8.59 hours** on master to **7.25 hours** for the PR.

Anyone can (and is encouraged to) reproduce the results by following this guide: https://gist.github.com/l0rinc/83d2bdfce378ad7396610095ceb7bed5

## Changing trends
The UTXO count and average block size have drastically increased in the past few years, providing a better overall picture of how Bitcoin behaves under real load.
[<img width="500" alt="image" src="https://github.com/user-attachments/assets/d257b864-d9de-4f9d-b61b-acc6835f384f" />](https://www.blockchain.com/explorer/charts/avg-block-size)
Profiling IBD, given these circumstances, revealed many new optimization opportunities.

## Similar efforts in the past years

There were many efforts to make sure Bitcoin Core remains performant in light of these new trends, a few recent notable examples include:
* https://github.com/bitcoin/bitcoin/pull/25325 - use specialized pool allocator for in-memory cache (~21% faster IBD)
* https://github.com/bitcoin/bitcoin/pull/28358 - allow the full UTXO set to fit into memory
* https://github.com/bitcoin/bitcoin/pull/28280#issuecomment-2273368482 - fine-grained in-memory cache eviction for pruned nodes (~30% IBD speedup on pruned nodes)
* https://github.com/bitcoin/bitcoin/pull/30039#issuecomment-2397229720 - reduce LevelDB writes, compactions and open files (~30% faster IBD for small in-memory cache)
* https://github.com/bitcoin/bitcoin/pull/31490, https://github.com/bitcoin/bitcoin/pull/30849, https://github.com/bitcoin/bitcoin/pull/30906 - refactors derisking/enabling follow-up optimizations
* https://github.com/bitcoin/bitcoin/pull/30326 - favor the happy path for cache misses (~2% IBD speedup)
* https://github.com/bitcoin/bitcoin/pull/30884 - Windows regression fix

## Reliable macro benchmarks

The measurements here were done on a high-end Intel i9-9900K CPU (8 cores/16 threads, 3.6GHz base, 5.0GHz boost), 64GB RAM, and a RAID configuration with multiple NVMe drives (total ~1.4TB fast storage), a dedicated Hetzner Auction box running latest Ubuntu. Sometimes a lower-end i7 was used with an HDD for comparison.

To make sure the setup reflected a real user's experience, we ran multiple full IBDs per commit (connecting to real nodes), until block 886'000 with a 5GiB in-memory cache where hyperfine was used to measure the final time (assuming normal distribution, stabilizing the final result via statistical methods), producing reliable results even when individual measurements varied (when hyperfine indicated that the measurements were all over the place we reran the whole benchmark).
To reduce the instability of headers synchronization and peer acquisition, we first started `bitcoind` until block 1, followed by the actual benchmarks until block 886'000.

The AutoFile batching with obfuscation batching changes (top 2 PRs) were measured my multiple people with different setting:
* @andrewtoth in https://github.com/bitcoin/bitcoin/pull/31144#issuecomment-2585403343 - 9% speedup with GCC
* @hodlinator in https://github.com/bitcoin/bitcoin/pull/31144#issuecomment-2588220255 - 11.9% speedup with GCC
* @mlori in https://github.com/bitcoin/bitcoin/pull/31144#issuecomment-2664077362 - 12% faster with GCC
* @Sjors in https://github.com/bitcoin/bitcoin/pull/31144#issuecomment-2575279833 - 3% speedup with Clang

Also note that there is a separate effort to [add a reliable macro-benchmarking suite](https://github.com/bitcoin/bitcoin/pull/31996) to track the performance of the most critical usecases end-to-end (including IBD, compact blocks, UTXO iteration) - still WIP.

## Current changes (in order of importance, reviews and reproducers are welcome):

Plotting the performance of the blocks from the produced `debug.log` files shows (from the last run, can differ slightly from the normalized average shown below) the effect of each commit:

<details>
<summary>debug.log visualizer</summary>

```python
import os
import sys
import re
from datetime import datetime
import matplotlib.pyplot as plt
import numpy as np


def process_log_files_and_plot(log_dir, output_file="block_height_progress.png"):
    if not os.path.exists(log_dir):
        print(f"Error: Directory '{log_dir}' does not exist", file=sys.stderr)
        return

    if not os.path.isdir(log_dir):
        print(f"Error: '{log_dir}' is not a directory", file=sys.stderr)
        return

    files = os.listdir(log_dir)
    debug_files = [f for f in files if f.startswith('debug-')]

    if not debug_files:
        print(f"Warning: No files starting with 'debug-' found in '{log_dir}'", file=sys.stderr)
        print(f"Files found: {files[:10]}{'...' if len(files) > 10 else ''}", file=sys.stderr)
        return

    timestamp_format = "%Y-%m-%dT%H:%M:%SZ"
    height_pattern = re.compile(r'UpdateTip:.+height=(\d+)')

    results = {}

    for filename in debug_files:
        filepath = os.path.join(log_dir, filename)
        if not os.path.isfile(filepath):
            print(f"Skipping '{filename}' as it's not a file", file=sys.stderr)
            continue

        print(f"Processing {filename}...", file=sys.stderr)

        update_tips = []
        first_timestamp = None
        line_count = 0
        tip_count = 0
        found_shutdown_done = False

        try:
            with open(filepath, 'r', errors='ignore') as file:
                for line_number, line in enumerate(file, 1):
                    line_count += 1
                    if line_count % 100000 == 0:
                        print(f"  Processed {line_count} lines, found {tip_count} UpdateTips...", file=sys.stderr)

                    if not found_shutdown_done:
                        if "Shutdown: done" in line:
                            found_shutdown_done = True
                            print(f"  Found 'Shutdown: done' at line {line_number}, starting to record",
                                  file=sys.stderr)
                        continue

                    if "UpdateTip:" not in line:
                        continue

                    try:
                        timestamp = datetime.strptime(line[:20], timestamp_format)

                        height_match = height_pattern.search(line)
                        if not height_match:
                            continue

                        height = int(height_match.group(1))

                        if first_timestamp is None:
                            first_timestamp = timestamp

                        elapsed_seconds = int((timestamp - first_timestamp).total_seconds())

                        update_tips.append((elapsed_seconds, height))
                        tip_count += 1

                    except ValueError:
                        continue
        except Exception as e:
            print(f"Error processing file '{filename}': {e}", file=sys.stderr)
            continue

        print(f"Finished processing {filename}: {line_count} lines, {tip_count} UpdateTips", file=sys.stderr)

        if update_tips:
            elapsed_time_to_height = {}
            for elapsed_time, height in update_tips:
                elapsed_time_to_height[elapsed_time] = height

            time_height_pairs = sorted(elapsed_time_to_height.items())

            results[filename[6:14]] = time_height_pairs
        else:
            print(f"Warning: No valid UpdateTip entries found in '{filename}'", file=sys.stderr)

    if not results:
        print("No results to plot. No valid UpdateTip entries found in any files.", file=sys.stderr)
        return

    print(f"Creating plot with data from {len(results)} files", file=sys.stderr)

    plt.figure(figsize=(12, 8))

    sorted_results = []
    for short_name, time_height_pairs in results.items():
        if time_height_pairs:
            last_time = time_height_pairs[-1][0] / 60 / 60
            sorted_results.append((short_name, last_time, time_height_pairs))

    sorted_results.sort(key=lambda x: x[1], reverse=True)

    colors = plt.cm.tab10(np.linspace(0, 1, len(sorted_results)))

    final_points = []

    for idx, (short_name, last_time_hours, time_height_pairs) in enumerate(sorted_results):
        times = [t for t, h in time_height_pairs]
        heights = [h for t, h in time_height_pairs]

        times_hours = [t / 60 / 60 for t in times]

        last_height = heights[-1] if heights else 0

        plt.plot(times_hours, heights, label=f"{short_name} ({last_time_hours:.2f}h)", color=colors[idx], linewidth=1)

        if times_hours and heights:
            final_points.append((last_time_hours, last_height, colors[idx]))

    for last_time, last_height, color in final_points:
        plt.axhline(y=last_height, color=color, linestyle='--', alpha=0.3)

        plt.axvline(x=last_time, color=color, linestyle='--', alpha=0.3)

    plt.title('Block Height Progress Over Time')
    plt.xlabel('Elapsed Time (hours)')
    plt.ylabel('Block Height')
    plt.grid(True, linestyle='--', alpha=0.7)

    plt.legend(loc='lower right')

    plt.tight_layout()

    plt.savefig(output_file, dpi=300)
    print(f"Plot saved to {output_file}", file=sys.stderr)

    with open(output_file.replace('.png', '.csv'), 'w') as f:
        for short_name, last_time, time_height_pairs in sorted_results:
            time_height_str = ",".join(f"{time}:{height}" for time, height in time_height_pairs)
            f.write(f"{short_name},{time_height_str}\n")

    plt.show()


if __name__ == "__main__":
    log_dir = sys.argv[1] if len(sys.argv) > 1 else "."
    output_file = sys.argv[2] if len(sys.argv) > 2 else "block_height_progress.png"
    print(f"Searching for log files in: {log_dir}", file=sys.stderr)
    process_log_files_and_plot(log_dir, output_file)
```

</details>

<img width="1000" alt="image" src="https://github.com/user-attachments/assets/9f13e4d7-6232-4a69-81b7-0b0d0cdbd9a2" />

### Baseline
Base commit was 88debb3e42.

<details>
<summary>8.59 hour IBD time</summary>

```bash
COMPILER=gcc COMMIT=88debb3e4297ef4ebc8966ffe599359bc7b231d0 ./build/src/bitcoind -datadir=/mnt/my_storage/BitcoinData -stopatheight=886000 -dbcache=5000 -blocksonly -printtoconsole=0
  Time (mean ± σ):     30932.610 s ± 156.891 s    [User: 58248.505 s, System: 2142.974 s]
  Range (min … max):   30821.671 s … 31043.549 s    2 runs
  ```

</details>

-----

* https://github.com/bitcoin/bitcoin/pull/31551

<details>
<summary>7.91 hour IBD time</summary>

```bash
COMPILER=gcc COMMIT=6a8ce46e32dae2ffef2a73d2314ca33a2039186e ./build/src/bitcoind -datadir=/mnt/my_storage/BitcoinData -stopatheight=886000 -dbcache=5000 -blocksonly -printtoconsole=0
  Time (mean ± σ):     28501.588 s ± 119.886 s    [User: 56419.060 s, System: 1833.126 s]
  Range (min … max):   28416.815 s … 28586.361 s    2 runs
```
</details>

We can serialize the blocks and undos to any `Stream` which implements the appropriate read/write methods.
`AutoFile` is one of these, writing the results "directly" to disk (through the OS file cache). Batching these in memory first and reading/writing these to disk is measurably faster (likely because of fewer native fread calls or less locking, as [observed](https://github.com/bitcoin/bitcoin/pull/28226#issuecomment-1666842501) by @martinus in a similar change).

Differential flame graphs indicate that the before/after speed change is because of fewer `AutoFile` read and writes:
![writes](https://github.com/user-attachments/assets/c5e45ddf-3a94-4c5b-999a-94a139d5776d)
![reads](https://github.com/user-attachments/assets/f11982ea-f2dc-4fc6-afce-faa132724658)

-----

* https://github.com/bitcoin/bitcoin/pull/31144

<details>
<summary>7.60 hour IBD time</summary>

```bash
COMPILER=gcc COMMIT=c5cc54d10187c9cb3a6cba8cc10f652b4f882e2a ./build/src/bitcoind -datadir=/mnt/my_storage/BitcoinData -stopatheight=886000 -dbcache=5000 -blocksonly -printtoconsole=0
  Time (mean ± σ):     27394.210 s ± 565.877 s    [User: 54902.315 s, System: 1891.951 s]
  Range (min … max):   26994.075 s … 27794.346 s    2 runs
```
</details>

Current block obfuscations are done byte-by-byte, this PR batches them to 64 bit primitives to speed up obfuscating bigger memory batches.
This is especially relevant after https://github.com/bitcoin/bitcoin/pull/31551 where we end up with bigger obfuscatable chunks.


-----

* https://github.com/bitcoin/bitcoin/pull/31645

<details>
<summary>7.50 hour IBD time</summary>

```bash
COMPILER=gcc COMMIT=9b4be912d20222b3b275ef056c1494a15ccde3f5 ./build/src/bitcoind -datadir=/mnt/my_storage/BitcoinData -stopatheight=886000 -dbcache=5000 -blocksonly -printtoconsole=0
  Time (mean ± σ):     27019.086 s ± 112.340 s    [User: 54927.344 s, System: 1652.376 s]
  Range (min … max):   26939.649 s … 27098.522 s    2 runs
```

</details>

The final UTXO set is [written to disk in batches](https://github.com/bitcoin/bitcoin/blob/master/src/txdb.cpp#L130-L133) to avoid a gigantic spike at flush time.
There is already a [-dbbatchsize](https://github.com/bitcoin/bitcoin/blob/master/src/init.cpp#L490) config option to change this value, this PR adjusts the default only.
By increasing the default batch size, we can reduce overhead from repeated compaction cycles, minimize constant overhead per batch, and achieve more sequential writes.

<img width="500" alt="image" src="https://github.com/user-attachments/assets/93ab3fff-385b-4093-8016-ab2a251cf92a" />

Note that this PR mainly optimizes a critical section of IBD - even if the effect on overall speed is small.


-----


* https://github.com/bitcoin/bitcoin/pull/31868

<details>
<summary>7.41 hour IBD time</summary>

```bash
COMPILER=gcc COMMIT=817d7ac0767a3984295aa3cf6c961dcc5f29d571 ./build/src/bitcoind -datadir=/mnt/my_storage/BitcoinData -stopatheight=886000 -dbcache=5000 -blocksonly -printtoconsole=0
  Time (mean ± σ):     26711.460 s ± 244.118 s    [User: 54654.348 s, System: 1652.087 s]
  Range (min … max):   26538.843 s … 26884.077 s    2 runs
```
</details>

This PR is draft since it relies on other PRs (e.g. https://github.com/bitcoin/bitcoin/pull/31519 and can be affected by the previous ones here).

The commits merge similar (de)serialization methods, and separates them internally with if constexpr - similarly to how it has been https://github.com/bitcoin/bitcoin/pull/28203. This enabled further SizeComputer optimizations as well.

Other than these, since single byte writes are used very often (used for every (u)int8_t or std::byte or bool and for every VarInt's first byte which is also needed for every (pre)Vector), it makes sense to avoid unnecessary generalized serialization infrastructure.

-----

* https://github.com/bitcoin/bitcoin/pull/31682

<details>
<summary>7.31 hour IBD time</summary>

```bash
COMPILER=gcc COMMIT=182745cec4c0baf2f3c8cff2f74f847eac3c4330 ./build/src/bitcoind -datadir=/mnt/my_storage/BitcoinData -stopatheight=886000 -dbcache=5000 -blocksonly -printtoconsole=0
  Time (mean ± σ):     26326.867 s ± 45.887 s    [User: 54367.156 s, System: 1619.348 s]
  Range (min … max):   26294.420 s … 26359.314 s    2 runs
```
</details>

`CheckBlock`'s latency is critical for efficiently validating correct inputs during transaction validation, including mempool acceptance and new block creation.

This PR improves performance and maintainability by introducing the following changes:
* Simplified checks for the most common cases (1 or 2 inputs - [70-90% of transactions have a single input](https://transactionfee.info/charts/transactions-1in)).
* Optimized the general case by replacing `std::set` with sorted `std::vector` for improved locality.
* Simplified Null prevout checks from linear to constant time.

-----

* https://github.com/bitcoin/bitcoin/pull/30442

<details>
<summary>7.25 hour IBD time</summary>

```bash
COMPILER=gcc COMMIT=47d377bd0bb88dae6b34553a7789400170e0ccf6 ./build/src/bitcoind -datadir=/mnt/my_storage/BitcoinData -stopatheight=886000 -dbcache=5000 -blocksonly -printtoconsole=0
  Time (mean ± σ):     26084.429 s ± 473.611 s    [User: 54310.780 s, System: 1815.967 s]
  Range (min … max):   25749.536 s … 26419.323 s    2 runs
```

</details>

The in-memory representation of the UTXO set uses [(salted) SipHash](https://github.com/bitcoin/bitcoin/blob/master/src/coins.h#L226) for avoiding key collision attacks.

Hashing a `uint256` key is done so often that a specialized optimization was extracted to [SipHashUint256Extra](https://github.com/bitcoin/bitcoin/blob/master/src/crypto/siphash.cpp#L135). The constant salting operations were already extracted in the [general case](https://github.com/bitcoin/bitcoin/blob/master/src/crypto/siphash.cpp#L20-L23), this PR adjusts the main specialization similarly.

-----

## Other similar efforts waiting for reviews or  revives (not included in this tracking PR):
* https://github.com/bitcoin/bitcoin/pull/31132 - pre-warms the in-memory cache on multiple threads (10% IBD speedup for small in-memory caches)
* https://github.com/bitcoin/bitcoin/pull/30611 - for very big in-memory caches make sure we still flush to disk regularly (no significant IBD speed change)
* https://github.com/bitcoin/bitcoin/pull/28945 - was meant to preallocate the memory of recreated caches (~6% IBD speedup for small caches)
* https://github.com/bitcoin/bitcoin/pull/31102 - was meant to try to evict entries selectively instead of dropping the whole cache when full

-----

This PR is meant to stay in draft (not meant to be merged directly), to continually change based on comments received here and in the PRs.
Comments and high-level discussions are welcome here - code reviews should rather be done in the individual PRs.