@ryanofsky 2025-03-12
Thanks for creating this. This should make it easier to navigate the other PRs and discuss the overall topic of IBD performance and benchmarking without needing to necessarily repeat it in the individual PRs.

Would be useful to have concept ACKs/NACKs here from others who know more about performance and benchmarking. But from from what I can tell the individual optimizations do not seem very complicated and seem like they should be justified.

---

One suggestion for the PR description above would be to directly link to the PRs comprising this change in the summary, maybe pointing out any where review should be focused. List seems to be:

- [ ] #31551 
- [ ] #31144
- [ ] #31645
- [ ] #31868
- [ ] #31682
- [ ] #30442