@l0rinc 2025-03-13
> with a simple "lower is better" comparison

I like the idea, updated the description and the code:
![image](https://github.com/user-attachments/assets/303426a5-a99f-4cb3-8552-a2def21d778b)

> Plotting the difference between the various proposed commits and the baseline (time_baseline[height] / time_commit_X[height], higher is better) might also be helpful? 

Something like this? Conceptually seems useful, but here I don't know how to interpret it, seems too far zoomed in
<img width="1000" alt="image" src="https://github.com/user-attachments/assets/14def0c8-241c-4e36-aa2d-554ce5fc4c15" />
Looks even funnier without the >400k cap:
<img width="1000" alt="image" src="https://github.com/user-attachments/assets/fb10ac16-54d8-4d59-bbaa-4fac9c4e0ac0" />