@ajtowns 2025-03-13
> Something like this? Conceptually seems useful, but here I don't know how to interpret it, seems too far zoomed in

Fair; that might work better with a time on the x-axis rather than height, something like:

```python
for time_commit in [time_commit_X, time_commit_Y, time_commit_Z]:
    for height in range(1,850000):
        y = time_baseline[height] / time_commit[height] # keep this one the same
        x = time_baseline[height]  # changed from x = height
        add_datapoint(x,y)
```