@l0rinc 2025-03-13
> time on the x-axis rather than height

If we do that we don't even need to filter out the first 400k blocks since they're so insignificant.
What do you think of this?

<img width="1000" alt="image" src="https://github.com/user-attachments/assets/1b3fb11a-f4f4-4277-99ba-decd2c60da6a" />