@jonatack 2025-03-13
Concept ACK, thank you for opening this.

I currently am in an environment of slow internet speed, where IBD is slower than 2 orders of magnitude worse than the times in the OP.

Opened #32051 today to address an issue I'm also seeing of very frequent disconnections+reconnections of trusted addnode peers during IBD.