@l0rinc 2025-07-20
We're making progress, https://github.com/bitcoin/bitcoin/pull/31144 was just merged! 🎉 
The next ones that need some love are:
- https://github.com/bitcoin/bitcoin/pull/32827
- https://github.com/bitcoin/bitcoin/pull/32497
- https://github.com/bitcoin/bitcoin/pull/32279