@pstratem 2025-08-01
We're calling CBlockHeader::GetHash hundreds of times for the same CBlockHeader object from ActivateBestChain.

I chased the calls with this commit and found all the results where from ProcessNewBlock ActivateBestChain

https://github.com/pstratem/bitcoin/commit/7de23309167d3f713f90cc552e1bd431464e4c49